Split RigidBodyTree.cpp into pieces
Split the various template methods and instantiations of the same
formerly in RigidBodyTree.cpp across several files. This previous
version of the file took nearly a minute to compile and required a
substantial amount of memory (3 GiB on my machine). The newly split
files compile in about 5s to 20s with worst memory requirements slightly
over (again, on my machine) 1 GiB. While this is actually worse overall
serial performance, it makes a significant impact on the ability of
machines with modest available memory to cope with multiple files being
compiled in parallel.

(In particular, the combination of pre-split RigidBodyTree.cpp,
RigidBodyTreeSDF.cpp, and RigidBodyTreeURDF.cpp is able to cripple my
machine, while it barely breaks a sweat even compiling most to all of
the newly split files in parallel. The latter two would benefit from
similar treatment.)