Move more setup logic to config.cmake
Move some additional logic to set up the build environment from
drake/CMakeLists.txt to cmake/config.cmake. This newly sets some stuff
for the superbuild, but since it is stuff that is not propagated to
externals, it should be harmless. This also exposes the option whether
or not to prefer to build shared libraries (which was already
propagated, but not shown unless the user had specified it manually).