Remove superfluous preprocessor symbol
Remove the logic in add_library_with_exports that adds the preprocessor
symbol <name>_STATIC_DEFINE to static libraries. Although this is
mentioned in the generated export headers, it is only needed when using
the same export header to build both a static and shared version of the
same library, which drake does not do.